pyproject.toml
[build-system]
requires = ["setuptools>=42.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "subtitle-translator-core"
version = "0.1.0"
description = "A FastAPI-based subtitle translation service"
authors = [
    {name = "Attila Kukel", email = "alitak@alitak.hu"},
]
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "fastapi[standard]==0.115.2",
    "python-dotenv==1.1.0",
    "sqlalchemy==2.0.41",
    "alembic==1.16.1",
    "yt-dlp==2025.5.22",
    "pydantic_settings=2.9.1"
]

[project.optional-dependencies]
win = ["python-magic-bin==0.4.14"]

[project.urls]
Homepage = "https://github.com/alitak/subtitle-translator-core"

[tool.setuptools]
include-package-data = true

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.setuptools.packages.find]
include = ["app"]
